+++ src/clj/l22/password.clj
@@ -23,7 +23,6 @@
     st/string]])
 
 (defn validate-password [params]
-  ;;(timbre/info "params" params)
   (first (st/validate params password-schema)))
 
 (defn password [{:keys [flash] :as request}]
@@ -34,15 +33,14 @@
   (if-let [errors (validate-password params)]
     (-> (response/found "/password")
         (assoc :flash (assoc params :errors errors)))
-    (let [user (db/get-user {:login (:login params)})
-          ret (hashers/verify (:password params) (:password user))]
-      (timbre/debug ret)
-      (if (:valid ret)
+    (let [user (db/get-user {:login (:login params)})]
+      (if (hashers/check (:password params) (:password user))
         (do
-          (db/update-user!
+          (db/update-password!
            (assoc (dissoc params :password)
                   :password (hashers/derive (:new-password params))))
-          (response/found "/"))
+          (-> (response/found "/")
+              (assoc :flash "password changed")))
         (layout/render nil "error.html"
                        {:status 404
                         :title "error"